+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -6,6 +6,9 @@
   * Updated to support a peer dependency of react-intl v2-v5
   * Updated FormattedMessage child function to join if given an array.
 
+* Fixed
+  * Fix to prevent modal from closing on pressing escape to close select.
+
 ## 3.28.0 - (August 4, 2020)
 
 * Changed