+++ packages/terra-application-navigation/CHANGELOG.md
@@ -6,6 +6,8 @@
   * Updated to support a peer dependency of react-intl v2-v5
   * Removed `intlShape` imports
 
+## 1.39.0 - (October 27, 2020)
+
 * Fixed
   * Corrected presentation of Help utility item in DrawerMenu when it is the only utility item present.