+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated to support a peer dependency of react-intl v2-v5
+  * Updated jest  due to changes in abstract modal.
 
 ## 4.3.0 - (October 20, 2020)